Update controller location and allow form to save properly
