+++ config/routes.rb
@@ -10,7 +10,7 @@ Spree::Core::Engine.add_routes do
   end
 
   resources :products, only: [] do
-    resources :product_reviews, only: [:index, :new, :create], controller: 'spree_product_reviews/spree/product_reviews' do
+    resources :product_reviews, only: [:index, :new, :create] do
     end
   end
 end
